This is an attempt to re-enter the TopDog Project into my GitHub repository.
GitHub is a universally used tool, but it is terrible. There are
so many holes between the web version, desktop version, and comand line version.

The best course of action that I have determined is to add repositories using
GitHub desktop which is logged into my account and then use the
command line commands to add the repository files.

TopDog is the first version of the Affiliate Website for TopDog
Trading. In reality it is a vechile for me to work out some issues
on how to use angularjs and css better.